Rename ooid to libooid
