+++ socorro/lib/libooid.py
