+++ webapp-django/crashstats/crashstats/tests/test_updatemissing.py
@@ -7,7 +7,7 @@ import os
 from crashstats.crashstats.models import MissingProcessedCrash
 from crashstats.crashstats.management.commands.updatemissing import Command
 from socorro.lib.libdatetime import utc_now
-from socorro.lib.ooid import create_new_ooid, date_from_ooid
+from socorro.lib.libooid import create_new_ooid, date_from_ooid
 
 
 TODAY = utc_now().strftime("%Y%m%d")
